Moving auth related code to the new shub.auth module
